bank/requirements.txt
django==4.0.4
psycopg2-binary==2.9.3
djangorestframework==3.14.0